refactor: extract mountReactComponent utility for improved component rendering
- Moved the `mountReactComponent` function to a new utility file for better organization and reusability.
- Updated the `initExcalidrawClientUI` function to utilize the new utility, adding a `className` option for the title container.
- Removed redundant code from the main UI file to streamline the component initialization process.
